Split email utilities by concern and extract dispatcher content helpers
utils/email.py mixed SMTP transport with branded-HTML template assembly
and social-share URL construction — three separate concerns. Replace the
single file with a utils/email package: smtp.py (pool, send, MIME,
is_email_configured), templates.py (load/render template, markdown→HTML),
and components.py (topic sections, TOC, share rows, topic colors, social
share URLs).

email_dispatcher.py carried per-subscriber topic-report filtering and
failure-JSON persistence in addition to its orchestration job. Extract
both into pipelines/node/email_subscriber_content.py so dispatch stays
focused on subscriber fetch → content build → wave send → stats, while
content filtering and failure persistence live together as their own
module.

Closes #49
