+++ utils/email/__init__.py
@@ -1,0 +1,7 @@
+"""Email package.
+
+Split by concern:
+- smtp: SMTP transport (connection pool, MIME, send).
+- templates: HTML template loading, markdown conversion, and top-level render.
+- components: HTML component builders (topic sections, TOC, share URLs/colors).
+"""
